src/pyproject.toml
[tool.poetry]
name = "ezstock"
version = "0.1.0"
description = ""
authors = ["desonglll <lindesong666@163.com>"]
readme = "README.md"

[tool.poetry.scripts]
django = "manage:main"

[tool.poetry.dependencies]
python = "^3.11"
asgiref = "3.7.2"
Babel = "2.14.0"
certifi = "2023.11.17"
charset-normalizer = "3.3.2"
click = "8.1.7"
colorama = "0.4.6"
Django = "5.0"
django-cors-headers = "4.3.1"
djangorestframework = "3.14.0"
djangorestframework-simplejwt = "5.3.1"
docutils = "0.20.1"
ghp-import = "2.1.0"
griffe = "0.39.1"
idna = "3.6"
Jinja2 = "3.1.3"
Markdown = "3.5.2"
MarkupSafe = "2.1.4"
mergedeep = "1.3.4"
mkdocs = "1.5.3"
mkdocs-autorefs = "0.5.0"
mkdocs-material = "9.5.4"
mkdocs-material-extensions = "1.3.1"
mkdocstrings = "0.24.0"
mkdocstrings-python = "1.8.0"
packaging = "23.2"
paginate = "0.5.6"
pathspec = "0.12.1"
pillow = "10.2.0"
platformdirs = "4.1.0"
Pygments = "2.17.2"
PyJWT = "2.8.0"
pymdown-extensions = "10.7"
python-dateutil = "2.8.2"
pytz = "2023.3.post1"
PyYAML = "6.0.1"
pyyaml_env_tag = "0.1"
regex = "2023.12.25"
requests = "2.31.0"
six = "1.16.0"
sqlparse = "0.4.4"
urllib3 = "2.1.0"
watchdog = "3.0.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
